+++ lib/index.tsx
@@ -1,0 +1,3 @@
+export * from "./blocks";
+export * from "./types";
+export * from "./utils";