+++ utils.ts
@@ -1,4 +1,0 @@
-import { custom } from '@rubriclab/chains'
-import type { ReactNode } from 'react'
-
-export const REACT_NODE = custom<ReactNode, 'ReactNode'>('ReactNode')